separate file for echo built
